chore: move tests of Deployment to single file
